feat: Migrate UFO_models to Python 3.6+
As FeynRules v2.3.35 produces Python 2 code only and modern MadGraph5_aMC@NLO
is Python 3 based (and is dropping support for Python 2 in December 2021
c.f. https://launchpad.net/mg5amcnlo/+announcement/22860 ) manually migrate all
UFO models to Python 3.6+ by using the 2to3 utility and pyupgrade with the
--py36-plus option.

c.f.:
- https://docs.python.org/3/library/2to3.html
- https://github.com/asottile/pyupgrade

The effects of this commit were carried out by:

* Running 2to3 in place on the codebase:

```
$ 2to3 --write .
```

* Manually undoing the effects of 2to3's use of list(dict.keys()) and
list(dict.items()) to reduce the amount of unnecessary diffs. dict.keys() and
dict.items() are iterables, and in the manner they are used in the UFO model
code it is as iterables and not as lists.

* Running pyupgrade with the --py36-plus option. For ease of use this was run
as a pre-commit hook, but the effects are equivalent to running the following

```
$ find -name '*.py' -print | xargs pyupgrade --py36-plus
```